Prevent future circular dependencies
